chore: update dependencies (including circle -> gha) (#47)
Test run:
https://github.com/dequelabs/dbjorge-test-repo/actions/runs/9604481533/job/26490058201

Notes on the updates:
- I blew away the whole lockfile and updated everything listed in
package.json to latest
- ...except eslint, which I left at major version 8 for compat with
typescript-eslint's current stable version
- ...and except a few dev deps that I dropped because we weren't using
them anyway (eslint+tslint stuff, sinon)
- The octokit update bumps octokit to a version that uses native fetch,
which nock can't mock, so I added a test seam that forces octokit to use
node-fetch when `NODE_ENV === 'test'`. The alternatives would be to use
a beta version of nock (didn't want to use a beta version of anything in
a repo we update so rarely) or to use a different mocking library
(`fetch-mock` was the best-looking option but it required more test
changes than I wanted to spend time on to swap it out)

I included a conversion from circle to GHA because I hit errors when
naively updating the circle config to node 20 and it was faster to
convert than to debug circle. None of the failed checks are relevant
anymore; once this merges we can disable circle in this repo and update
the branch policies to refer to GHA.

## Reviewer checks

**Required fields, to be filled out by PR reviewer(s)**

- [ ] Follows the commit message policy, appropriate for next version
- [ ] Code is reviewed for security